Fixed CoreData Factory methods
