+++ The Movie DB Client/App/ManagedObjectContextFactory.swift
@@ -6,15 +6,10 @@
 //  Copyright © 2018 Amadeu Cavalcante Filho. All rights reserved.
 //
 
-import UIKit
 import CoreData
 
 final class ManagedObjectContextFactory {
     static func make() -> NSManagedObjectContext? {
-        if let appDelegate = UIApplication.shared.delegate as? AppDelegate {
-            appDelegate.persistentContainer.viewContext.mergePolicy = NSMergeByPropertyObjectTrumpMergePolicy
-            return appDelegate.persistentContainer.viewContext
-        }
-        return nil
+        return PersistentContainerFactory.make().viewContext
     }
 }
