+++ The Movie DB Client/App/ManagedObjectModelFactory.swift
@@ -6,14 +6,10 @@
 //  Copyright © 2018 Amadeu Cavalcante Filho. All rights reserved.
 //
 
-import UIKit
 import CoreData
 
 final class ManagedObjectModelFactory {
     static func make() -> NSManagedObjectModel? {
-        if let appDelegate = UIApplication.shared.delegate as? AppDelegate {
-            return appDelegate.persistentContainer.managedObjectModel
-        }
-        return nil
+        return PersistentContainerFactory.make().managedObjectModel
     }
 }
